Format string fixes:
f_count is long
ok kettenis@

Change esm_sensor_map.arg to int
ok deraadt@

+++ sys/arch/i386/i386/esm.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: esm.c,v 1.52 2010/07/02 01:35:13 tedu Exp $ */
+/*	$OpenBSD: esm.c,v 1.53 2012/10/17 22:31:59 deraadt Exp $ */
 
 /*
  * Copyright (c) 2005 Jordan Hargrave <jordan@openbsd.org>
@@ -87,7 +87,7 @@ enum sensor_type esm_typemap[] = {
 
 struct esm_sensor_map {
 	enum esm_sensor_type	type;
-	long			arg;
+	int			arg;
 	const char		*name;
 };
